refactor(simplify): move most functions to their specific repo
