+++ lib/get-entity-name.js
