+++ lib/listr.js
@@ -1,4 +1,4 @@
-import { logToTaskOutput, formatLogMessageOneLine } from '../utils/logging'
+import { logToTaskOutput, formatLogMessageOneLine } from './logging'
 
 // Set up log emitter listening from SDK, proper error catching and throwing of SDK errors
 export function wrapTask (func) {
